# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bestellungs-Übersicht | API-Backend in Laravel",
        "description": "Diese Web-App verwaltet Bestellungen und die dazugehörenden Produkte und Geschäfte.\n *     Dieses Projekt wurde im Laufe einer IPA erstellt.",
        "contact": {
            "email": "julien.raedler@twofold.swiss"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "User Register",
                "description": "User Register here",
                "operationId": "Register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "3|W4pFsiVbdGz9C0f0ZAaDtMSLnTxNMex7j5WM2V7F"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The route could not be found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "User Login",
                "description": "Login User Here",
                "operationId": "authLogin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "3|W4pFsiVbdGz9C0f0ZAaDtMSLnTxNMex7j5WM2V7F"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The route could not be found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders": {
            "get": {
                "tags": [
                    "order"
                ],
                "summary": "Get all orders",
                "description": "Returns list of orders",
                "operationId": "getOrders",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "orders": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OrderResource"
                                            },
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/orderProducts": {
            "get": {
                "tags": [
                    "orderProduct"
                ],
                "summary": "Get all ordered products",
                "description": "Returns list of ordered products",
                "operationId": "getOrderProducts",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "orderProducts": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OrderProductResource"
                                            },
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "OrderProductResource": {
                "properties": {
                    "product": {
                        "$ref": "#/components/schemas/ProductResource"
                    },
                    "recieved_at": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "null"
                    },
                    "volume": {
                        "type": "integer",
                        "example": 25
                    },
                    "price": {
                        "type": "integer",
                        "example": 100
                    },
                    "price_total": {
                        "type": "integer",
                        "example": 2500
                    }
                },
                "type": "object",
                "xml": {
                    "name": "OrderProducts"
                }
            },
            "OrderResource": {
                "properties": {
                    "ordered_at": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "1991-06-14"
                    },
                    "user": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Hans"
                    },
                    "completed": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "ordered_products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderProductResource"
                        },
                        "collectionFormat": "multi"
                    },
                    "order_price": {
                        "type": "integer",
                        "maxLength": 32,
                        "example": "2500"
                    },
                    "note": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "This is a note about taking notes off things"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Orders"
                }
            },
            "ProductResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "shop": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "Digitec"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "USB cable"
                    },
                    "image": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "977630ae369389ee38ca474af9e3d30e.png"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Products"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}